requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
boto3==1.16.11
boto3-stubs==1.16.11.0
botocore==1.19.11
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
ConfigArgParse==1.0
dataclasses-json==0.5.1
dateparser==0.7.2
idna==2.8
iniconfig==1.0.1
isodate==0.6.0
jmespath==0.10.0
markdown-it-py==git+https://github.com/executablebooks/markdown-it-py.git@v1.0.0b1
marshmallow==3.7.0
marshmallow-enum==1.5.1
mdit-py-plugins==0.2.6
mypy-boto3-cloudfront==1.16.11.0
mypy-boto3-route53==1.16.11.0
mypy-boto3-s3==1.16.11.0
mypy-extensions==0.4.3
packaging==20.4
pathvalidate==0.29.0
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pyformance==0.4
pyparsing==2.4.5
pytest==6.1.1
python-dateutil==2.8.1
pytz==2019.3
pyyaml-5.4.1
rdflib==5.0.0
rdflib-jsonld==0.5.0
regex==2019.12.19
requests==2.22.0
s3transfer==0.3.3
six==1.13.0
stringcase==1.2.0
toml==0.10.1
tqdm==4.45.0
typing-extensions==3.7.4.2
typing-inspect==0.6.0
tzlocal==2.0.0
urllib3==1.25.7
